refactor: move private hook to public
